applied some changes to assure that code is compilable using Intel compiler on Windows
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@863 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/application/common/ScreenOutput.cpp
@@ -17,7 +17,7 @@ ScreenOutput::ScreenOutput()
 	first_run = true;
 }
 
-void ScreenOutput::update(int progress, std::string status)
+void ScreenOutput::update(size_t progress, std::string status)
 {
 	int i;
 	GetConsoleScreenBufferInfo(hCon,&info);
